requirements.txt
cryptography
ipcalc
netifaces
six
systemd-python
PyYAML